+++ vsd/CHANGELOG.md
@@ -23,8 +23,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - `save`
-    - `--set-cookie` and `--user-agent` flags removed.
+    - `--no-certificate-checks`, `--set-cookie` and `--user-agent` flags removed.
     - Default retries changed from 15 to 10.
+    - Now `--output` flag works with `--parse` flag
 - `capture`
     - Feature re-worked.